a bug about nowait flag has been fixed.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist-Python/branches/RELENG_1_1@472 3ceb7c45-664a-4a73-a5f3-5d010aa2b7c3

+++ OpenRTM_aist/PeriodicExecutionContext.py
@@ -685,7 +685,7 @@ class PeriodicExecutionContext(OpenRTM_aist.ExecutionContextBase,
     if rate > 0.0:
       self._profile.rate = rate
       self._period.set_time(1.0/rate)
-      if long(self._period.toDouble()) == 0:
+      if self._period.toDouble() == 0.0:
         self._nowait = True
 
       for comp in self._comps:
